+++ src/serializers.ts
@@ -1,34 +1,5 @@
 import getImageUrl from 'getImageUrl'
 
-export interface Serializers {
-  // Common overrides
-  types: {
-    block: any
-    image: any
-  }
-  marks: any
-
-  // Less common overrides
-  list: any
-  listItem: any
-
-  block: any
-  span: any
-  hardBreak: any
-
-  // Container element
-  container: any
-
-  // When we can't resolve the mark properly, use this renderer as the container
-  markFallback: any
-
-  // Allow overriding text renderer, but leave undefined to just use plain strings by default
-  text?: any
-
-  // Empty nodes (React uses null, hyperscript with empty strings)
-  empty: string
-}
-
 export default function serializers(h, serializerOpts?: any) {
   const serializeOptions = serializerOpts || {useDashedStyles: false}
